backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
email-validator==2.1.0
pdf2image==1.16.3
Pillow==10.1.0
python-pptx==0.6.23
aiofiles==23.2.1
python-dotenv==1.0.0
jinja2==3.1.2
